don't create $HADOOP_HOME/bin/jsvc link if jsvc is already there as a binary.

+++ bin/hbase-ec2-init-remote.sh
@@ -127,7 +127,7 @@ fi
 # Security setup
 # all servers need krb5 libraries
 yum -y install krb5-libs jakarta-commons-daemon-jsvc
-ln -s /usr/bin/jsvc $HADOOP_HOME/bin
+[ -f $HADOOP_HOME/bin/jsvc ] || ln -s /usr/bin/jsvc $HADOOP_HOME/bin
 adduser hadoop
 if [ "$IS_MASTER" = "true" ]; then
   yum -y install krb5-server
